deleted test files and migrated the code to HamburgOttensen component (/maps)
