+++ src/components/Dashboard.vue
@@ -3,21 +3,21 @@
 		<div>
 			<app-header></app-header>
 		</div>
-		<app-test></app-test>
+		<app-hamburg-ottensen></app-hamburg-ottensen>
 	</div>
 </template>
 
 <script>
 
 import Header from "./Header.vue"
-import Test from "./Test.vue"
+import HamburgOttensen from "./maps/HamburgOttensen"
 
 
 export default {
 	name: "appDashboard",
 	components: {
 		"appHeader": Header,
-		"appTest": Test
+		"appHamburgOttensen": HamburgOttensen
 	}
 }
