+++ src/components/maps/HamburgOttensen.vue
@@ -1,39 +1,100 @@
-<!-- <template>
-	<div class="indoormap">
-		<v-app id="inspire">
-		    <div class="text-xs-center">
-		      <v-menu offset-y>
-		        <v-btn color="primary" dark slot="activator"></v-btn>
-		        <v-list>
-		          <v-list-tile v-for="option in options" :key="option.title" @click="">
-		            <v-list-tile-title>{{ option.title }}</v-list-tile-title>
-		          </v-list-tile>
-		        </v-list>
-		      </v-menu>
-		    </div>
-  		</v-app>		
-</div>	
+<template>
+  <div>
+    <div class="dropdown">
+          <input type="button" 
+          class="dropdown-toggle" 
+          v-for="desk in deskList" 
+          :value="desk.id"
+          @click="handler(desk.id)"
+          />                            
+      <div>        
+         <ul class="dropdown-menu" v-if="showList" v-click-outside="hideList"> Check In
+            <li v-for="option in options">
+              <input type="button" @click="showModal" :value="option.item">
+            </li>
+          </ul>
+      </div>
+      <modal name="search">
+        <app-search :action="checkIn"></app-search>
+      </modal>
+    </div>
+</div>
+</body>  
+
 </template>
 
 <script>
 
+import axios from "axios"
+import Search from "../Search.vue"
+
 export default {
-	name: "HamburgOttensen",
-	data() {
-		return {
-			options: [
-		      { title: 'Check Someone In' },
-		      { title: 'Report An Issue' },
-			]
-    	}	
-  	} 
+
+  name: "HamburgOttensen",
+   data() {
+    return {
+	    deskList: [],
+      selectedDesk: null,
+      showList:false,
+      options: [
+        {
+          item: "Myself"
+        },
+        {
+          item: "Someone Else"
+        },
+        {
+          item: "Report An Issue"
+        }
+      ]
+    }   
+  },
+  components: {
+    "appSearch": Search
+  },
+  created() {
+    axios.get("https://s2-move.firebaseio.com/desks.json")
+      .then(res => {
+          const data = res.data
+          const desks = []
+          for (let key in data) {
+            const desk = data[key]
+            desk.id = key
+            desks.push(desk)
+          }
+          this.deskList = desks
+            // console.log(desks)
+          })
+      .catch(error => console.log(error))
+  },
+  methods: {
+    selectDesk(deskId) {
+      this.selectedDesk = deskId;
+      console.log(this.selectedDesk);
+    },
+    handler(deskId) {
+      this.selectDesk(deskId)
+      this.showList = !this.showList
+    },
+    hideList() {
+      this.showList = false;
+    },
+    showModal () {
+      this.$modal.show('search');
+    },
+    hideModal () {
+      this.$modal.hide('search');
+    },
+    checkIn(selectedUser) {
+      console.log("I'm the check in")
+    }
+  }
 }
 
 </script>
 
 <style lang="css">
 
-@import "../../../styles/css/hamburgottensen.component.css"
+@import "../../../styles/css/ottensen.component.css"
 
 </style>
- -->
